pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wiazor-identity"
version = "1.0.0"
description = "id.wiazor.com — Authentication & Organization Service for DTKT Platform"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "DTKT Architecture Team" },
]

dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "asyncpg>=0.29.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "nats-py>=2.7.0",
    "httpx>=0.27.0",
    "cryptography>=42.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/identity"]
